Created git repository
This project started a little bit ealier but as if grew in size and
complexity I decided to make a git repository to help manage it and view
changes, and to future proof it by helping with version control. Current
version is 0.3 and thats for basically no reason at all.