Added node pages, updated version, readme, all files. V 0.7
There is a new feature of having click-able cores on the main queue
pages, and then it will take you to a page for the node itself
displaying core info and jobs for that specific node. I also added the
nice feature of a compelted setup script and better automation script.
After configuring the setup script, it should a breeeze to install. I
also aded the python daemon as thats the most important piece after all.
The idea of the project changed a tad bit, as I decreed its out of Alpha
stages and into Beta with this version 0.7. There will most likely be
quick changes and I am about to set this up on an openstack instance,
fingers crossed.