Bug fixes for beta-2
Fixed a lot of bugs and uglies since I was lazy and used the latest
commit to transfer files before testing. I added the forgotten field for
status in the pending jobs, adjusted pending header to reflect new info
I found, moved the links on the Queue pages, and fixed the error which
would make the daemon exit because someone's job was in error state.