Beta-3. Added new method of setup and grabbing files.
Added a new way to setup and have the whole system works. To better
protect my sanity, Dr. Hampton suggest going a route of grabbing files
using wget from my own www directory, so I don't need to hardcode my
password into anything. So, instead of using wget I decided to go with
curl as it can easily overwrite files in the process of downloading them
which is what we want after-all. Also changed the way the system dies
when coming across a bad pending job status (learn something new about
uge everyday). Now it will write to a log what the status was that
caused the death, instead of mysteriously dying with exit code 20.
Updated the readme to reflect new additions. For a cosmetic appeal, I
added a faveicon.ico to be displayed on all of the pages. This is the
same one found on crc.nd.edu, so if you're an outsider using this repo
I'd suggest changing it to what your organization is.