Fixed pending job status error in daemon, updated readme.
There was an issue with the way queue_mapd.py handled a currently
unknown status for a pending job, which im theory is now handled and it
should from now on correctly write that status to a log and die. The
readme was updated to reflect the need to specify the paths in
queue_mapd.py itself before running it.