Fixed fault in daemon.
There was a previously unknown status for a job waiting in a queue,
which kept causing the daemon to die. Thus was the purpose of the log
addition: to find out what that status was. Turns out when a job is
being pulled from pending into running, it gets a status of 'Rq', I did
not know that. That has now been added to the possible status(es?) so
that bug is now caught and in theory the daemon should run indefinitely
now.