v1.0.
Created memory filter/view for the available queues/hostgroups. In memory view each node is shown as once rectangle and it is displayed as a certain color depending on how much memory is being used by that node. Introduced more logging capabilities for errors which cause daemon death. Prepping for the addition of more host-groups to be displayed. Change the way the daemon finds information pertaining to the UGE, it now uses 'qstat -F' vs 'qstat -f'. This way the daemon can find a lot more information. Added the memory information to the node pages, nodes are still a link in the memory view/filter.