Added date to logging, made page_down.html
Figured the date and time may come in handy for the logging in
queue_mapd.log. Also added the capability to write to a log from the
create_memory_html function. Created page_down.html to be set as the
index.html for /var/www/html in times when configuring or something
breaks.