Created Development Branch. Beginning of new features.
The new branch will house all of the new features which are not ready to be implemented in the master yet.
There are a few new ideas, see README for details. With the beginning of the new branch there is a new way of
obtaining and organizing nodes found from UGE, as each hostgroup will now be pulled. Each hostgroup will
also have a list of nodes associated with it. The hostgroups will be split into the core usage and memeory usage
for easy organizing and finding through links. Only the updated curl_setup.sh is ready for testing, and ready
for a new name as it does more than merely curl a few files.