Daemon now cleans up its mess after generating text files for each HG. Ditched Queue-tracking.
Older versions keep track of which queue is being viewed, Long or Debug, that
idea is now gone. The flow is moving towards the host-groups now, and so the
queue doesn't really matter as the hostgroup will be a better method to use
to keep track of what we're looking at anyway.